// web/src/prompt-studio/styles/tokens.d.ts
declare const colors: {
  primary: {
    DEFAULT: string;
    hover: string;
    container: string;
    containerDark: string;
    onPrimary: string;
    onContainer: string;
    onContainerDark: string;
    selection: string;
    editorBackground: string;
  };
  surface: {
    DEFAULT: string;
    variant: string;
    onSurface: string;
    onVariant: string;
    dark: string;
    variantDark: string;
    onSurfaceDark: string;
    onVariantDark: string;
    container: string;
    containerDark: string;
    containerLow: string;
    containerLowDark: string;
    containerHigh: string;
    containerHighDark: string;
    containerHighest: string;
    containerHighestDark: string;
  };
  background: {
    DEFAULT: string;
    dark: string;
  };
  border: {
    DEFAULT: string;
    dark: string;
  };
  error: {
    DEFAULT: string;
    container: string;
    containerDark: string;
    onContainer: string;
    onContainerDark: string;
  };
  text: {
    light: {
      primary: string;
      secondary: string;
      muted: string;
    };
    dark: {
      primary: string;
      secondary: string;
      muted: string;
    };
  };
  scrollbar: {
    thumb: string;
    thumbDark: string;
    thumbHover: string;
  };
};

export { colors };